+++ internal/services/logic/logic_app_standard_resource_test.go
@@ -1001,6 +1001,7 @@ provider "azurerm" {
 resource "azurerm_storage_share" "custom" {
   name                 = "customshare"
   storage_account_name = azurerm_storage_account.test.name
+  quota                = 1
 }
 
 resource "azurerm_logic_app_standard" "test" {